rename githubutils to githubapi
