+++ pkg/githubapi/githubapi.go
@@ -1,4 +1,4 @@
-package githubutils
+package githubapi
 
 import (
 	"encoding/json"
